+++ packages/livedoc-mocha/_src/app/reporter/ColorTheme.ts
@@ -27,5 +27,7 @@ export interface ColorTheme {
     dataTableHeader: Chalk;
     dataTable: Chalk;
     docString: Chalk;
-    valuePlaceholders: Chalk
+    valuePlaceholders: Chalk;
+
+    summaryHeader: Chalk;
 }